+++ frontend/mock/welcomePageCards.json
@@ -1,0 +1,73 @@
+[
+  {
+    "commentText":
+    "Классно быть партнером Feed Market. Я развиваю компанию во всех направлениях: от продаж своих товаров на маркетплейсе до ведения своих пунктов выдачи и помощи другим предпринимателям в выходе на площадку Feed Market и другие маркетплейсы.",
+    "userName": "Юзернейм Юзернеймович",
+    "userDescription":
+    "Генеральный директор ООО «Компания нейм», селлер в категориях «Категория нейм», «Категория нейм 2»",
+    "userImage": "/../public/welcomePage/Ellipse 632.png"
+  },
+  {
+    "commentText": "Feed Market приносит нам большую часть оборота с маркетплейсов.",
+    "userName": "Юзернейм Юзернеймович",
+    "userDescription":
+    "Генеральный директор ООО «Компания нейм», селлер в категориях «Категория нейм», «Категория нейм 2»",
+    "userImage": "/../public/welcomePage/Ellipse 632.png"
+  },
+  {
+    "commentText": "Feed Market приносит нам большую часть оборота с маркетплейсов.",
+    "userName": "Юзернейм Юзернеймович",
+    "userDescription":
+    "Генеральный директор ООО «Компания нейм», селлер в категориях «Категория нейм», «Категория нейм 2»",
+    "userImage": "/../public/welcomePage/Ellipse 632.png"
+  },
+  {
+    "commentText": "Feed Market приносит нам большую часть оборота с маркетплейсов.",
+    "userName": "Юзернейм Юзернеймович",
+    "userDescription":
+    "Генеральный директор ООО «Компания нейм», селлер в категориях «Категория нейм», «Категория нейм 2»",
+    "userImage": "/../public/welcomePage/Ellipse 632.png"
+  },
+  {
+    "commentText": "Feed Market приносит нам большую часть оборота с маркетплейсов.",
+    "userName": "Юзернейм Юзернеймович",
+    "userDescription":
+    "Генеральный директор ООО «Компания нейм», селлер в категориях «Категория нейм», «Категория нейм 2»",
+    "userImage": "/../public/welcomePage/Ellipse 632.png"
+  },
+  {
+    "commentText": "Feed Market приносит нам большую часть оборота с маркетплейсов.",
+    "userName": "Юзернейм Юзернеймович",
+    "userDescription":
+    "Генеральный директор ООО «Компания нейм», селлер в категориях «Категория нейм», «Категория нейм 2»",
+    "userImage": "/../public/welcomePage/Ellipse 632.png"
+  },
+  {
+    "commentText": "Feed Market приносит нам большую часть оборота с маркетплейсов.",
+    "userName": "Юзернейм Юзернеймович",
+    "userDescription":
+    "Генеральный директор ООО «Компания нейм», селлер в категориях «Категория нейм», «Категория нейм 2»",
+    "userImage": "/../public/welcomePage/Ellipse 632.png"
+  },
+  {
+    "commentText": "Feed Market приносит нам большую часть оборота с маркетплейсов.",
+    "userName": "Юзернейм Юзернеймович",
+    "userDescription":
+    "Генеральный директор ООО «Компания нейм», селлер в категориях «Категория нейм», «Категория нейм 2»",
+    "userImage": "/../public/welcomePage/Ellipse 632.png"
+  },
+  {
+    "commentText": "Feed Market приносит нам большую часть оборота с маркетплейсов.",
+    "userName": "Юзернейм Юзернеймович",
+    "userDescription":
+    "Генеральный директор ООО «Компания нейм», селлер в категориях «Категория нейм», «Категория нейм 2»",
+    "userImage": "/../public/welcomePage/Ellipse 632.png"
+  },
+  {
+    "commentText": "Feed Market приносит нам большую часть оборота с маркетплейсов.",
+    "userName": "Юзернейм Юзернеймович",
+    "userDescription":
+    "Генеральный директор ООО «Компания нейм», селлер в категориях «Категория нейм», «Категория нейм 2»",
+    "userImage": "/../public/welcomePage/Ellipse 632.png"
+  }
+]
